requirements.txt
duckdb
pyarrow
websockets
streamlit
pandas
numpy
requests
cryptography
python-dotenv
